Add files via upload
This is presentation material for the competition which is written in korean. 
My part was p-median modeling whereas there is other model using MCLP algorithm made by my teammate. 
I was a leader of the team and has contributed to the project a lot especially by comparing and interpreting the results of two methods.